refactor: move setDisplayName to settings lib folder
