+++ lib/settings/setDisplayName.ts
